Scrolling and basic XML specification - Created the basic definition for XML based View inflation      - Can be found under xml-layout-spec.xml      - Will be implemented in the future as a separate module as an addition to this - Started work on the ScrollLayout      - Basically done, just need to implement support globally for scrolls and consuming scrolls - Cleaned up some other misc. stuff - Added debug(str: String) method to debug various messages.

+++ src/com/asdev/libjam/md/drawable/RippleDrawable.kt
@@ -106,14 +106,14 @@ class RippleDrawable(
 
         rippleX -= 10
 
-        g.clipRect(rippleX.toInt(),
+        g.clipRect(rippleX,
                 y.toInt(),
                 10,
                 h.toInt())
 
         // draw a circle at the end of each ripple
         g.fillOval(
-                (rippleX).toInt(),
+                rippleX,
                 y.toInt() - h.toInt() / 2,
                 h.toInt(),
                 h.toInt() * 2
